Add cancel event on tools
